refactored blast code for clarity
